Added dynamic player dust effect
Player now kicks up dust when they run
There's a small burst on initial movement, direction change and stopping, followed by a smaller trail if you continue running.

The particles are children of the player model

Created "Player" sort layer and tag to get the particles to render behind the player

Created materials folder for dust material (if you want to change the colour create a new dust material so we can compare between the options easily)

Need to get it to not emit when jumping as it should only emit when on the ground

I think it looks pretty!